+++ lib/flutter_simple_country_picker.dart
@@ -52,6 +52,6 @@ export 'src/localization/country_localizations.dart' show CountryLocalizations;
 export 'src/model/country.dart';
 export 'src/theme/country_picker_theme.dart';
 export 'src/util/country_input_formatter.dart';
-export 'src/widget/country_phone_input.dart';
+export 'src/widget/country_phone_input.dart' hide CountryPhoneInput$Extended;
 export 'src/widget/country_scope.dart' show CountryScope;
 export 'src/widget/show_country_picker.dart' show showCountryPicker;